fix(SessionHandler): Keys within sessionhandler are now case-insensitive
- added a min > sec converer and the reverse to calculator.
- checIfNumber now parses doubles instead of int (mistake from leftover code.)
- updated diary

+++ src/main/java/se/iths/utility/ScannerWrapper.java
@@ -74,7 +74,7 @@ public class ScannerWrapper {
             }
 
             try {
-                userInput = Integer.parseInt(input);
+                userInput = Double.parseDouble(input);
                 if (userInput < 0) {
                     System.err.print("Please enter a positive number: ");
                 } else {
